blind warning fix in wxUmaskChanger
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@29637 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/filefn.h
@@ -560,7 +560,7 @@ public:
     // wxFileConfig::Flush() for example
     wxUmaskChanger(int umaskNew)
     {
-        m_umaskOld = umaskNew == -1 ? -1 : umask((mode_t)umaskNew);
+        m_umaskOld = umaskNew == -1 ? -1 : (int)umask((mode_t)umaskNew);
     }
 
     ~wxUmaskChanger()
